fix(csa-todo): stop attesting placeholder TODO.md so draft plans don't false-report [PLAN TAMPERED] (#1669)
`TodoManager::create` wrote a SHA-256 attestation over the throwaway
placeholder TODO.md ("# TODO: <title>\n\n## Goal\n\n## Tasks\n\n- [ ] \n").
That placeholder is ALWAYS overwritten with the real plan: the mktd workflow
(Step 13) writes TODO.md directly via `printf >`, then `csa todo save`
re-attests. Any plan whose content was written before a subsequent attest
therefore reported `Mismatch` -> "[PLAN TAMPERED]" on `csa todo show`, even for
pristine, never-tampered drafts. This trained operators to ignore the banner --
destroying the signal value of a genuine tamper warning -- and in
`pipeline_plan_context` silently dropped the plan's spec/criteria from pipeline
context.

Root-cause fix: `create_inner` no longer attests the placeholder. A freshly
created, not-yet-finalized plan now reports `Missing` (benign, no banner), while
a genuine post-attestation edit still reports `Mismatch`. `ensure_attestation`
(run by `csa todo save`), `attest`, and `write_todo_md` remain the points that
establish/refresh the real attestation -- verified end to end: create -> direct
`printf >` write -> `csa todo save` now yields a clean `Valid`, no banner.

- csa-todo: remove the placeholder attestation in `create_inner`; the two
  manual-tamper tests now attest a baseline first (tamper is only meaningful
  after a real attestation exists); add lifecycle + un-attested-draft regression
  tests.
- cli-sub-agent: extend the `[PLAN TAMPERED]` banner with a `csa todo attest`
  remediation hint (issue suggestion #2). Extract todo_cmd.rs's inline test
  module to todo_cmd_tests.rs -- todo_cmd.rs was a pre-existing 8877-token
  monolith and touching it would otherwise trip the token-budget gate; the
  extraction follows the established `#[path = "*_tests.rs"]` convention and
  drops todo_cmd.rs to 7766 tokens.

Fixes #1669
